feat(editor): add EditorProvider with ref support
Add an EditorProvider component that wraps tiptap's EditorProvider and
supports a ref exposing EmailEditorRef methods (getEmail, getEmailHTML,
getEmailText, getJSON, editor).

This gives users of the composable EditorProvider API the same ref
capabilities that the batteries-included EmailEditor component provides.

- Extract shared buildRef, RefBridge, and ReadyBridge into
  editor-provider/ref-bridge.tsx
- Create EditorProvider wrapper with forwardRef in
  editor-provider/editor-provider.tsx
- Export EditorProvider and EditorProviderProps from package root
- Refactor EmailEditor to reuse shared ref utilities
