+++ packages/editor/src/editor-provider/index.ts
@@ -1,0 +1,2 @@
+export { EditorProvider, type EditorProviderProps } from './editor-provider';
+export { type EmailEditorRef } from './ref-bridge';
