+++ packages/editor/src/index.ts
@@ -3,3 +3,8 @@ export {
   type EmailEditorProps,
   type EmailEditorRef,
 } from './email-editor';
+
+export {
+  EditorProvider,
+  type EditorProviderProps,
+} from './editor-provider';
